test: move appendStackTrace regression to capture-stack-trace.test.js
error-gc-test.test.js has unrelated tests that time out under the ASAN
debug build; moving to a file that runs clean there.
